requirements.txt
opencv-python
requests
pyzbar
tqdm
names
PyYAML